requirements.txt
pyserial>=3.5
numpy>=1.24
h5py>=3.8
matplotlib>=3.7
scikit-learn>=1.3
tensorflow>=2.14
scipy>=1.11
